third_party/WebKit/LayoutTests/inspector/tracing/timeline-network/timeline-network-resource.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/timeline-test.js"></script>
<script>

var scriptUrl = "../../../http/tests/inspector/tracing/resources/timeline-network-resource.js";

function performActions()
{
    var promise = new Promise((fulfill) => window.timelineNetworkResourceEvaluated = fulfill);
    var script = document.createElement("script");
    script.src = scriptUrl;
    document.body.appendChild(script);
    return promise;
}

function test()
{
    var requestId;
    var scriptUrl = "timeline-network-resource.js";

    InspectorTest.invokeAsyncWithTimeline("performActions", finish);

    function finish()
    {
        var model = InspectorTest.timelineModel();
        model.mainThreadEvents().forEach(event => {
            if (event.name === TimelineModel.TimelineModel.RecordType.ResourceSendRequest)
                printSend(event);
            else if (event.name === TimelineModel.TimelineModel.RecordType.ResourceReceiveResponse)
                printReceive(event);
            else if (event.name === TimelineModel.TimelineModel.RecordType.ResourceFinish)
                printFinish(event);
        });
        InspectorTest.completeTest();
    }

    function printEvent(event)
    {
        InspectorTest.addResult("");
        InspectorTest.printTraceEventProperties(event);
        InspectorTest.addResult(`Text details for ${event.name}: ` + Timeline.TimelineUIUtils.buildDetailsTextForTraceEvent(event));
    }

    function printSend(event)
    {
        printEvent(event);
        var data = event.args["data"];
        requestId = data.requestId;
        if (data.url === undefined)
            InspectorTest.addResult("* No 'url' property in record");
        else if (data.url.indexOf(scriptUrl) === -1)
            InspectorTest.addResult("* Didn't find URL: " + scriptUrl);
    }

    function printReceive(event)
    {
        printEvent(event);
        var data = event.args["data"];
        if (requestId !== data.requestId)
            InspectorTest.addResult("Didn't find matching requestId: " + requestId);
        if (data.statusCode !== 0)
            InspectorTest.addResult("Response received status: " + data.statusCode);
    }

    function printFinish(event)
    {
        printEvent(event);
        var data = event.args["data"];
        if (requestId !== data.requestId)
            InspectorTest.addResult("Didn't find matching requestId: " + requestId);
        if (data.didFail)
            InspectorTest.addResult("Request failed.");
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests the Timeline API instrumentation of a network resource load
</p>
</body>
</html>
