third_party/WebKit/LayoutTests/inspector/tracing/timeline-network/timeline-network-resource-expected.txt

Tests the Timeline API instrumentation of a network resource load

Script resource loaded

ResourceSendRequest Properties:
{
    data : {
        frame : <string>
        priority : "Low"
        requestId : <string>
        requestMethod : "GET"
        stackTrace : <object>
        url : .../inspector/tracing/resources/timeline-network-resource.js
    }
    endTime : <number>
    frameId : <string>
    stackTrace : <object>
    startTime : <number>
    type : "ResourceSendRequest"
}
Text details for ResourceSendRequest: timeline-network-resource.js

ResourceReceiveResponse Properties:
{
    data : {
        encodedDataLength : <number>
        frame : <string>
        fromCache : false
        fromServiceWorker : false
        mimeType : <string>
        requestId : <string>
        statusCode : 0
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    type : "ResourceReceiveResponse"
}
Text details for ResourceReceiveResponse: timeline-network-resource.js

ResourceFinish Properties:
{
    data : {
        decodedBodyLength : 223
        didFail : false
        encodedDataLength : <number>
        finishTime : <number>
        requestId : <string>
    }
    endTime : <number>
    startTime : <number>
    type : "ResourceFinish"
}
Text details for ResourceFinish: timeline-network-resource.js
